Changes after CR: refactor ShoppingCartTable and BaseTable add: ShoppingCartHeader enum and ShoppingCartTableRowAndColumn
